fixed some more tests
