+++ tests/shard2/test_cost_function_integration.py
@@ -192,6 +192,7 @@ def test_pendulum(control_type, integration_rule, objective, phase_dynamics):
                         controls[:, i : i + 2].reshape((-1, 1)),  # Controls
                         [],  # Parameters
                         [],  # Algebraic states
+                        [],  # Dynamics constants
                         ocp.nlp[0].J[0].weight,  # Weight
                         [],  # Target
                     )
